removed duplicates and re-use tests classes from cluster-core
